backend/requirements.txt
Flask>=2.0
Flask-SQLAlchemy>=3.0
Flask-Migrate>=4.0
psycopg2-binary
python-dotenv
PyJWT
Werkzeug
urllib